pyproject.toml
[project]
name = "tanghouse"
version = "0.1.0"
description = ""
authors = [
    {name = "hshi"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "google-genai (>=1.45.0,<2.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "tqdm (>=4.67.1,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

package-mode = false
[tool.poetry.group.dev.dependencies]
ipykernel = "^6.30.1"

